requirements.txt
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
google-api-core==1.26.3
google-auth==1.29.0
google-cloud-pubsub==2.4.1
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
idna==2.10
iniconfig==1.1.1
libcst==0.3.18
mypy-extensions==0.4.3
packaging==20.9
pluggy==0.13.1
proto-plus==1.18.1
protobuf==3.15.8
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
